pyproject.toml
[project]
name = "dotclaude"
version = "0.1.0"
description = "Personal Claude Code configuration with skills, commands, agents, and hooks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.57.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
# All optional dependencies for extended functionality. Install with:
# uv pip install -e ".[all]" or install specific groups as needed.
# These can be lazily installed only when the corresponding features are used.

# Claude usage metrics tracking (scripts/claude-usage.py)
usage = ["httpx>=0.25.0"]

# YouTube content extraction and analysis (skills/youtube-content/scripts/)
youtube = [
    "youtube-transcript-api>=0.6.1",
    "yt-dlp>=2024.1.0",
]

# Image generation with multiple AI providers (skills/image-gen/scripts/)
images = [
    "openai>=1.3.0",
    "google-genai>=0.3.0",
    "fal-client>=0.2.0",
    "requests>=2.31.0",
]

# MCP (Model Context Protocol) builder and Anthropic integration (skills/mcp-builder/)
mcp = [
    "anthropic>=0.39.0",
    "mcp>=1.1.0",
]

# Configuration validation and schema checking (hooks/validators, .github/scripts/)
config = ["jsonschema>=4.20.0"]

# Development and testing convenience: install all optional dependencies
all = [
    "httpx>=0.25.0",
    "youtube-transcript-api>=0.6.1",
    "yt-dlp>=2024.1.0",
    "openai>=1.3.0",
    "google-genai>=0.3.0",
    "fal-client>=0.2.0",
    "requests>=2.31.0",
    "anthropic>=0.39.0",
    "mcp>=1.1.0",
    "jsonschema>=4.20.0",
]
